+++ lib/record/index.d.ts
@@ -1,14 +1,15 @@
-import { MixableConstructor } from '../object-plus';
+import { CollectionConstructor } from '../collection';
+import { TheType } from '../object-plus';
+import { Transactional } from '../transactions';
 import { Infer } from './attrDef';
 import { Record } from './record';
-import { CollectionConstructor } from '../collection';
 export * from './attrDef';
 export * from './metatypes';
 export { Record };
 export declare type InferAttrs<A extends object> = {
     [K in keyof A]: Infer<A[K]>;
 };
-export interface RecordConstructor<A> extends MixableConstructor {
+export interface RecordConstructor<A> extends TheType<typeof Transactional> {
     new (attrs?: Partial<A>, options?: object): Record & A;
     prototype: Record;
     Collection: CollectionConstructor<Record & A>;