+++ lib/relations/from.d.ts
@@ -1,2 +1,2 @@
 import { Record } from '../record';
-export declare function from(path: string, T?: typeof Record): any;
+export declare function from<R extends typeof Record>(path: string, T?: R): any;