+++ src/record/index.ts
@@ -1,10 +1,10 @@
-import { predefine, tools, MixableConstructor, define } from '../object-plus';
+import { CollectionConstructor } from '../collection';
+import { define, predefine, TheType, tools } from '../object-plus';
 import { Transactional } from '../transactions';
-import { Infer, createSharedTypeSpec, type } from './attrDef';
+import { createSharedTypeSpec, Infer, type } from './attrDef';
 import { SharedType } from './metatypes';
 import { createAttributesMixin } from './mixin';
 import { Record, RecordDefinition } from './record';
-import { CollectionConstructor } from '../collection';
 
 export * from './attrDef';
 export * from './metatypes';
@@ -16,7 +16,7 @@ export type InferAttrs<A extends object> = {
     [K in keyof A]: Infer<A[K]>
 };
 
-export interface RecordConstructor<A> extends MixableConstructor {
+export interface RecordConstructor<A> extends TheType<typeof Transactional> {
     new ( attrs? : Partial<A>, options? : object ) : Record & A
     prototype : Record
     Collection : CollectionConstructor<Record & A>